fix rtrouted crash is seen with an invalid DML query ( OFW-5723 )
[Issue]
When a dmcli query is made with a large number of dots, rtrouted crashes.
For example: dmcli eRT getv Device.NAT.X_CISCO_COM_DMZ.InternalIP.....................................................................

The reason is that the tokenizeExpression function splits the datamodel path based on dots and writes into the workTokens array without checking the number of elements being written.
This causes a buffer overflow if the number of dots is more than the size of the workTokens array (32).

[Solution]
Add some checks in tokenizeExpression to avoid overflowing the workTokens and workBuffer arrays.

[Testing performed]
Run 'dmcli eRT getv Device.NAT.X_CISCO_COM_DMZ.InternalIP.....................................................................'
Verify that rtrouted doesn't crash
Run 'rbuscli get' with a path that's more than 512 characters in length
Verify that rtrouted doesn't crash